rename: Payment, ReservationDay 위치 변경
