+++ src/main/java/christmas/domain/benefit/Badge.java
@@ -1,6 +1,6 @@
 package christmas.domain.benefit;
 
-import christmas.domain.Payment;
+import christmas.domain.order.Payment;
 import christmas.view.output.DomainMessage;
 
 public class Badge {
