+++ src/main/java/christmas/domain/benefit/Benefit.java
@@ -2,10 +2,9 @@ package christmas.domain.benefit;
 
 import static java.lang.String.format;
 
-import christmas.domain.Payment;
+import christmas.domain.order.Payment;
 import christmas.view.output.DomainMessage;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Benefit {
     private final int benefitAmount;
