+++ src/main/java/christmas/domain/benefit/SpecialDayPolicy.java
@@ -2,8 +2,8 @@ package christmas.domain.benefit;
 
 import static java.lang.String.format;
 
-import christmas.domain.Payment;
-import christmas.domain.ReservationDay;
+import christmas.domain.order.Payment;
+import christmas.domain.order.ReservationDay;
 import christmas.view.output.DomainMessage;
 import java.util.Arrays;
 import java.util.List;
