+++ src/main/java/christmas/domain/benefit/WeekendPolicy.java
@@ -2,11 +2,10 @@ package christmas.domain.benefit;
 
 import static java.lang.String.format;
 
-import christmas.domain.Payment;
-import christmas.domain.ReservationDay;
+import christmas.domain.order.Payment;
+import christmas.domain.order.ReservationDay;
 import christmas.view.output.DomainMessage;
 import java.time.DayOfWeek;
-import java.time.LocalDate;
 
 public class WeekendPolicy extends DayOfWeekPolicy {
     private static final String MAIN_DISH = "MainDish";
