+++ src/main/java/christmas/domain/order/ReservationDay.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 import static java.lang.String.format;
