+++ src/test/java/christmas/domain/benefit/BenefitTest.java
@@ -3,8 +3,8 @@ package christmas.domain.benefit;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import christmas.domain.Payment;
-import christmas.domain.ReservationDay;
+import christmas.domain.order.Payment;
+import christmas.domain.order.ReservationDay;
 import christmas.domain.order.OrderHistory;
 import java.util.ArrayList;
 import java.util.List;
